LayoutTests/http/tests/security/contentSecurityPolicy/block-all-mixed-content/duplicate-directive.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="block-all-mixed-content; block-all-mixed-content">
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<p>This tests that we emit a console warning when block-all-mixed-content is listed more than once.</p>
</body>
</html>
