Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/block-all-mixed-content/duplicate-directive-expected.txt

CONSOLE MESSAGE: Ignoring duplicate Content-Security-Policy directive 'block-all-mixed-content'.

This tests that we emit a console warning when block-all-mixed-content is listed more than once.
